requirements.txt
Django
djangorestframework
django-ratelimit
redis
django-dotenv